Added scene loading
Renamed the load function in main to Setup; this function should contain everything to set up the game. e.g. adding scenes to the SceneManager
Added LoadScene method to the SceneManager. This method takes an integer and loads the scene with that index
Only one scene can be loaded at the same time
Moved Main file out of the engine filter, so it is easily accesible