Added Init to SceneManager and Scene
Added Init method to the SceneManager and Scene (which calls the Init function of the gameobjects)
The Init method of the components is no longer called on creation of the component. They are now called in the init method of the gameobject, thus at the end of scene creation.
Moved Transform creation on a GameObject from the Init method to the constructor