Added templated DataCommand
Added a DataCommand class that derives from Command and is templated. The template typename is the variable that gets stored in the command.
Eg. a datacommand<float> has a datamember of type float
Modified existing commands to use DataCommand where appropriate
Modified InputManager to check if commands are DataCommands, and if so pass data to them